ABW-285 - Data class param renamed.
