+++ app/src/main/java/com/babylon/wallet/android/presentation/model/AccountUi.kt
@@ -6,5 +6,5 @@ data class AccountUi(
     val hash: String,
     val amount: String,
     val currencySymbol: String,
-    val accountAssets: List<String>
+    val assets: List<String>
 )
